# openapi.json
{
"openapi": "3.1.0",
"info": {
"title": "NEXUS PM Agent API",
"version": "0.1.0"
},
"paths": {
...
"/team-members": {
"get": {
"tags": [
"team"
],
"summary": "List Members",
"operationId": "list_members_team_members_get",
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"items": {
"additionalProperties": true,
"type": "object"
},
"type": "array",
"title": "Response List Members Team Members Get"
}
}
}
}
}
},
"post": {
"tags": [
"team"
],
"summary": "Create Member",
"operationId": "create_member_team_members_post",
"requestBody": {
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/MemberBody"
}
}
},
"required": true
},
"responses": {
"201": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"additionalProperties": true,
"type": "object",
"title": "Response Create Member Team Members Post"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/team-members/{member_id}": {
"patch": {
"tags": [
"team"
],
"summary": "Patch Member",
"operationId": "patch_member_team_members__member_id__patch",
"parameters": [
{
"name": "member_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Member Id"
}
}
],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/MemberPatch"
}
}
}
},
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "object",
"additionalProperties": true,
"title": "Response Patch Member Team Members  Member Id  Patch"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/stakeholders": {
"get": {
"tags": [
"team"
],
"summary": "List Stakeholders",
"operationId": "list_stakeholders_stakeholders_get",
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"items": {
"additionalProperties": true,
"type": "object"
},
"type": "array",
"title": "Response List Stakeholders Stakeholders Get"
}
}
}
}
}
},
"post": {
"tags": [
"team"
],
"summary": "Create Stakeholder",
"operationId": "create_stakeholder_stakeholders_post",
"requestBody": {
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/StakeholderBody"
}
}
},
"required": true
},
"responses": {
"201": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"additionalProperties": true,
"type": "object",
"title": "Response Create Stakeholder Stakeholders Post"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
"/stakeholders/{stakeholder_id}": {
"delete": {
"tags": [
"team"
],
"summary": "Delete Stakeholder",
"operationId": "delete_stakeholder_stakeholders__stakeholder_id__delete",
"parameters": [
{
"name": "stakeholder_id",
"in": "path",
"required": true,
"schema": {
"type": "integer",
"title": "Stakeholder Id"
}
}
],
"responses": {
"200": {
"description": "Successful Response",
"content": {
"application/json": {
"schema": {
"type": "object",
"additionalProperties": true,
"title": "Response Delete Stakeholder Stakeholders  Stakeholder Id  Delete"
}
}
}
},
"422": {
"description": "Validation Error",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/HTTPValidationError"
}
}
}
}
}
}
},
...
},
"components": {
"schemas": {
...
"HTTPValidationError": {
"properties": {
"detail": {
"items": {
"$ref": "#/components/schemas/ValidationError"
},
"type": "array",
"title": "Detail"
}
},
"type": "object",
"title": "HTTPValidationError"
},
...
"MemberBody": {
"properties": {
"name": {
"type": "string",
"title": "Name"
},
"role": {
"type": "string",
"title": "Role"
},
"skill_tags": {
"items": {
"type": "string"
},
"type": "array",
"title": "Skill Tags",
"default": []
},
"capacity_hrs": {
"type": "number",
"title": "Capacity Hrs",
"default": 40
}
},
"type": "object",
"required": [
"name",
"role"
],
"title": "MemberBody"
},
"MemberPatch": {
"properties": {
"name": {
"anyOf": [
{
"type": "string"
},
{
"type": "null"
}
],
"title": "Name"
},
"role": {
"anyOf": [
{
"type": "string"
},
{
"type": "null"
}
],
"title": "Role"
},
"skill_tags": {
"anyOf": [
{
"items": {
"type": "string"
},
"type": "array"
},
{
"type": "null"
}
],
"title": "Skill Tags"
},
"capacity_hrs": {
"anyOf": [
{
"type": "number"
},
{
"type": "null"
}
],
"title": "Capacity Hrs"
},
"is_active": {
"anyOf": [
{
"type": "boolean"
},
{
"type": "null"
}
],
"title": "Is Active"
},
"user_id": {
"anyOf": [
{
"type": "integer"
},
{
"type": "null"
}
],
"title": "User Id"
}
},
"type": "object",
"title": "MemberPatch"
},
...
"StakeholderBody": {
"properties": {
"name": {
"type": "string",
"title": "Name"
},
"audience_type": {
"type": "string",
"title": "Audience Type"
},
"email": {
"anyOf": [
{
"type": "string"
},
{
"type": "null"
}
],
"title": "Email"
},
"project_id": {
"anyOf": [
{
"type": "integer"
},
{
"type": "null"
}
],
"title": "Project Id"
}
},
"type": "object",
"required": [
"name",
"audience_type"
],
"title": "StakeholderBody"
},
...
"ValidationError": {
"properties": {
"loc": {
"items": {
"anyOf": [
{
"type": "string"
},
{
"type": "integer"
}
]
},
"type": "array",
"title": "Location"
},
"msg": {
"type": "string",
"title": "Message"
},
"type": {
"type": "string",
"title": "Error Type"
},
"input": {
"title": "Input"
},
"ctx": {
"type": "object",
"title": "Context"
}
},
"type": "object",
"required": [
"loc",
"msg",
"type"
],
"title": "ValidationError"
}
}
}
}